Added classes and their respective animators.

+++ Assets/Scripts/Projectiles/ProjectilesTypeData.cs
@@ -6,6 +6,6 @@ public class ProjectileTypeData : ScriptableObject
     public string typeName;
     public GameObject prefab;      // Prefab for the projectile visuals
     public float speed;            // Speed of the projectile
-    public float damage;           // Damage dealt by this projectile
+    public int damage;           // Damage dealt by this projectile
    // public float lifetime;         // How long the projectile exists
 }
